+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/drm/OfflineLicenseHelper.java
@@ -34,7 +34,6 @@ import androidx.media3.exoplayer.drm.DrmSession.DrmSessionException;
 import androidx.media3.exoplayer.source.MediaSource.MediaPeriodId;
 import com.google.common.util.concurrent.SettableFuture;
 import java.util.Map;
-import java.util.UUID;
 import java.util.concurrent.ExecutionException;
 import org.checkerframework.checker.nullness.compatqual.NullableType;
 
@@ -125,25 +124,6 @@ public final class OfflineLicenseHelper {
             .build(
                 new HttpMediaDrmCallback(
                     defaultLicenseUrl, forceDefaultLicenseUrl, dataSourceFactory)),
-        eventDispatcher);
-  }
-
-  /**
-   * @deprecated Use {@link #OfflineLicenseHelper(DefaultDrmSessionManager,
-   *     DrmSessionEventListener.EventDispatcher)} instead.
-   */
-  @Deprecated
-  public OfflineLicenseHelper(
-      UUID uuid,
-      ExoMediaDrm.Provider mediaDrmProvider,
-      MediaDrmCallback callback,
-      @Nullable Map<String, String> optionalKeyRequestParameters,
-      DrmSessionEventListener.EventDispatcher eventDispatcher) {
-    this(
-        new DefaultDrmSessionManager.Builder()
-            .setUuidAndExoMediaDrmProvider(uuid, mediaDrmProvider)
-            .setKeyRequestParameters(optionalKeyRequestParameters)
-            .build(callback),
         eventDispatcher);
   }